+++ release-notes.md
@@ -1,3 +1,5 @@
+- `InstancesSet.api_client` is now public
+
 v 0.11.0
 ========